feat: UI for info and success screens
