+++ features/exploredash/src/main/res/layout/explore_dash_main_info.xml
@@ -150,7 +150,7 @@
         app:layout_constraintBottom_toBottomOf="parent"
         android:layout_marginBottom="48dp"
         android:layout_marginHorizontal="16dp"
-        android:text="@string/continue_text"
+        android:text="@string/button_continue"
         />
 
     <androidx.constraintlayout.widget.Guideline
