+++ wallet/AndroidManifest.xml
@@ -92,12 +92,12 @@
         </activity>
 
         <activity
-            android:name="de.schildbach.wallet.ui.VerifySeedActivity"
+            android:name="de.schildbach.wallet.ui.verify.VerifySeedActivity"
             android:screenOrientation="portrait"
             android:theme="@style/My.Theme.ChildActivity" />
 
         <activity
-            android:name="de.schildbach.wallet.ui.ViewSeedActivity"
+            android:name="de.schildbach.wallet.ui.verify.ViewSeedActivity"
             android:screenOrientation="portrait"
             android:theme="@style/My.Theme.ChildActivity" />
