+++ wallet/res/drawable/ic_backup_info.xml
@@ -1,0 +1,54 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:aapt="http://schemas.android.com/aapt"
+    android:width="80dp"
+    android:height="80dp"
+    android:viewportWidth="80"
+    android:viewportHeight="80">
+  <path
+      android:pathData="M40,0L40,0A40,40 0,0 1,80 40L80,40A40,40 0,0 1,40 80L40,80A40,40 0,0 1,0 40L0,40A40,40 0,0 1,40 0z">
+    <aapt:attr name="android:fillColor">
+      <gradient 
+          android:startX="40"
+          android:startY="0"
+          android:endX="40"
+          android:endY="80"
+          android:type="linear">
+        <item android:offset="0" android:color="#FF78C4F5"/>
+        <item android:offset="1" android:color="#FF31A3EF"/>
+      </gradient>
+    </aapt:attr>
+  </path>
+  <path
+      android:pathData="M39.718,60.273C42.493,60.273 45.098,59.742 47.535,58.679C49.986,57.629 52.144,56.171 54.011,54.304C55.877,52.437 57.342,50.285 58.405,47.848C59.468,45.398 60,42.779 60,39.992C60,37.218 59.468,34.612 58.405,32.175C57.342,29.725 55.877,27.566 54.011,25.7C52.144,23.833 49.986,22.375 47.535,21.324C45.098,20.261 42.493,19.73 39.718,19.73C36.944,19.73 34.339,20.255 31.901,21.305C29.464,22.342 27.312,23.788 25.445,25.641C23.579,27.495 22.114,29.641 21.051,32.078C19.988,34.515 19.45,37.114 19.437,39.875H23.89C23.903,37.685 24.318,35.643 25.134,33.75C25.951,31.844 27.085,30.179 28.537,28.753C29.989,27.327 31.668,26.212 33.574,25.408C35.479,24.591 37.528,24.183 39.718,24.183C41.909,24.183 43.958,24.591 45.863,25.408C47.782,26.225 49.46,27.359 50.9,28.811C52.351,30.25 53.486,31.929 54.302,33.847C55.132,35.753 55.547,37.801 55.547,39.992C55.547,42.196 55.132,44.257 54.302,46.176C53.486,48.081 52.351,49.76 50.9,51.212C49.46,52.664 47.782,53.798 45.863,54.615C43.958,55.432 41.909,55.833 39.718,55.821C38.046,55.821 36.458,55.581 34.954,55.101C33.464,54.621 32.083,53.941 30.812,53.059C29.555,52.165 28.44,51.115 27.468,49.909C27.014,49.417 26.489,49.112 25.893,48.995C25.296,48.866 24.739,49.008 24.22,49.423C23.728,49.799 23.456,50.298 23.404,50.92C23.365,51.53 23.598,52.132 24.104,52.729C25.271,54.22 26.677,55.535 28.323,56.676C29.983,57.817 31.785,58.699 33.729,59.321C35.687,59.956 37.683,60.273 39.718,60.273ZM17.123,35.889C16.591,35.889 16.19,36.012 15.917,36.258C15.658,36.505 15.535,36.822 15.548,37.211C15.561,37.587 15.71,37.97 15.995,38.359L20.973,45.165C21.31,45.618 21.712,45.845 22.179,45.845C22.645,45.832 23.041,45.605 23.365,45.165L28.362,38.339C28.635,37.95 28.771,37.568 28.771,37.192C28.784,36.816 28.66,36.505 28.401,36.258C28.142,36.012 27.747,35.889 27.215,35.889H17.123Z"
+      android:fillColor="#ffffff"/>
+  <path
+      android:pathData="M32.661,34.895H41.37C41.668,34.895 41.917,34.787 42.118,34.572C42.318,34.356 42.418,34.107 42.418,33.824C42.418,33.526 42.318,33.272 42.118,33.061C41.923,32.85 41.674,32.745 41.37,32.745H32.661C32.363,32.745 32.114,32.853 31.914,33.069C31.713,33.279 31.613,33.531 31.613,33.824C31.613,34.102 31.713,34.351 31.914,34.572C32.119,34.787 32.369,34.895 32.661,34.895ZM44.769,34.895H47.636C47.929,34.895 48.176,34.787 48.376,34.572C48.582,34.356 48.684,34.107 48.684,33.824C48.684,33.526 48.584,33.272 48.384,33.061C48.183,32.85 47.934,32.745 47.636,32.745H44.769C44.481,32.745 44.235,32.853 44.029,33.069C43.829,33.279 43.729,33.531 43.729,33.824C43.729,34.102 43.831,34.351 44.037,34.572C44.243,34.787 44.486,34.895 44.769,34.895ZM32.661,38.942H33.609C33.902,38.942 34.149,38.834 34.349,38.618C34.55,38.397 34.65,38.145 34.65,37.863C34.65,37.565 34.55,37.31 34.349,37.1C34.154,36.889 33.907,36.784 33.609,36.784H32.661C32.369,36.784 32.119,36.891 31.914,37.107C31.713,37.323 31.613,37.575 31.613,37.863C31.613,38.14 31.713,38.389 31.914,38.61C32.119,38.831 32.369,38.942 32.661,38.942ZM37.008,38.942H40.623C40.916,38.942 41.162,38.834 41.363,38.618C41.568,38.397 41.671,38.145 41.671,37.863C41.671,37.565 41.571,37.31 41.37,37.1C41.17,36.889 40.921,36.784 40.623,36.784H37.008C36.715,36.784 36.469,36.891 36.268,37.107C36.068,37.323 35.968,37.575 35.968,37.863C35.968,38.14 36.068,38.389 36.268,38.61C36.474,38.831 36.72,38.942 37.008,38.942ZM44.021,38.942H47.636C47.934,38.942 48.183,38.834 48.384,38.618C48.584,38.397 48.684,38.145 48.684,37.863C48.684,37.565 48.584,37.31 48.384,37.1C48.183,36.889 47.934,36.784 47.636,36.784H44.021C43.729,36.784 43.482,36.891 43.282,37.107C43.081,37.323 42.981,37.575 42.981,37.863C42.981,38.14 43.081,38.389 43.282,38.61C43.487,38.831 43.734,38.942 44.021,38.942ZM32.661,42.988H35.528C35.811,42.988 36.055,42.877 36.261,42.656C36.466,42.435 36.569,42.186 36.569,41.909C36.569,41.621 36.466,41.372 36.261,41.161C36.06,40.945 35.816,40.838 35.528,40.838H32.661C32.363,40.838 32.114,40.943 31.914,41.153C31.713,41.364 31.613,41.616 31.613,41.909C31.613,42.196 31.713,42.448 31.914,42.664C32.119,42.88 32.369,42.988 32.661,42.988ZM38.927,42.988H47.636C47.929,42.988 48.176,42.877 48.376,42.656C48.582,42.435 48.684,42.186 48.684,41.909C48.684,41.621 48.584,41.372 48.384,41.161C48.183,40.945 47.934,40.838 47.636,40.838H38.927C38.629,40.838 38.38,40.943 38.18,41.153C37.979,41.364 37.879,41.616 37.879,41.909C37.879,42.196 37.979,42.448 38.18,42.664C38.38,42.88 38.629,42.988 38.927,42.988ZM32.661,47.026H40.361C40.654,47.026 40.9,46.918 41.101,46.702C41.306,46.487 41.409,46.237 41.409,45.955C41.409,45.657 41.306,45.403 41.101,45.192C40.9,44.976 40.654,44.868 40.361,44.868H32.661C32.363,44.868 32.114,44.976 31.914,45.192C31.713,45.408 31.613,45.662 31.613,45.955C31.613,46.232 31.713,46.479 31.914,46.695C32.119,46.916 32.369,47.026 32.661,47.026Z"
+      android:strokeAlpha="0.5"
+      android:fillAlpha="0.5">
+    <aapt:attr name="android:fillColor">
+      <gradient 
+          android:startX="40.149"
+          android:startY="32.745"
+          android:endX="40.149"
+          android:endY="47.026"
+          android:type="linear">
+        <item android:offset="0" android:color="#FFFFFFFF"/>
+        <item android:offset="1" android:color="#FFF2F2F2"/>
+      </gradient>
+    </aapt:attr>
+  </path>
+  <path
+      android:pathData="M32.661,34.895H41.37C41.668,34.895 41.917,34.787 42.118,34.572C42.318,34.356 42.418,34.107 42.418,33.824C42.418,33.526 42.318,33.272 42.118,33.061C41.923,32.85 41.674,32.745 41.37,32.745H32.661C32.363,32.745 32.114,32.853 31.914,33.069C31.713,33.279 31.613,33.531 31.613,33.824C31.613,34.102 31.713,34.351 31.914,34.572C32.119,34.787 32.369,34.895 32.661,34.895ZM44.021,38.942H47.636C47.934,38.942 48.183,38.834 48.384,38.618C48.584,38.397 48.684,38.145 48.684,37.863C48.684,37.565 48.584,37.31 48.384,37.1C48.183,36.889 47.934,36.784 47.636,36.784H44.021C43.729,36.784 43.482,36.891 43.282,37.107C43.081,37.323 42.981,37.575 42.981,37.863C42.981,38.14 43.081,38.389 43.282,38.61C43.487,38.831 43.734,38.942 44.021,38.942ZM32.661,42.988H35.528C35.811,42.988 36.055,42.877 36.261,42.656C36.466,42.435 36.569,42.186 36.569,41.909C36.569,41.621 36.466,41.372 36.261,41.161C36.06,40.945 35.816,40.838 35.528,40.838H32.661C32.363,40.838 32.114,40.943 31.914,41.153C31.713,41.364 31.613,41.616 31.613,41.909C31.613,42.196 31.713,42.448 31.914,42.664C32.119,42.88 32.369,42.988 32.661,42.988ZM32.661,47.026H40.361C40.654,47.026 40.9,46.918 41.101,46.702C41.306,46.487 41.409,46.237 41.409,45.955C41.409,45.657 41.306,45.403 41.101,45.192C40.9,44.976 40.654,44.868 40.361,44.868H32.661C32.363,44.868 32.114,44.976 31.914,45.192C31.713,45.408 31.613,45.662 31.613,45.955C31.613,46.232 31.713,46.479 31.914,46.695C32.119,46.916 32.369,47.026 32.661,47.026Z">
+    <aapt:attr name="android:fillColor">
+      <gradient 
+          android:startX="40.149"
+          android:startY="32.745"
+          android:endX="40.149"
+          android:endY="47.026"
+          android:type="linear">
+        <item android:offset="0" android:color="#FFFFFFFF"/>
+        <item android:offset="1" android:color="#FFF2F2F2"/>
+      </gradient>
+    </aapt:attr>
+  </path>
+</vector>
