+++ wallet/res/drawable/ic_does_not_store.xml
@@ -1,0 +1,47 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:aapt="http://schemas.android.com/aapt"
+    android:width="34dp"
+    android:height="34dp"
+    android:viewportWidth="34"
+    android:viewportHeight="34">
+  <path
+      android:pathData="M17,0L17,0A17,17 0,0 1,34 17L34,17A17,17 0,0 1,17 34L17,34A17,17 0,0 1,0 17L0,17A17,17 0,0 1,17 0z"
+      android:fillColor="#78C4F5"/>
+  <path
+      android:pathData="M9.72,12.598H17.647C17.919,12.598 18.145,12.5 18.328,12.303C18.51,12.107 18.601,11.88 18.601,11.623C18.601,11.351 18.51,11.12 18.328,10.928C18.15,10.736 17.923,10.641 17.647,10.641H9.72C9.448,10.641 9.222,10.739 9.039,10.935C8.857,11.127 8.766,11.356 8.766,11.623C8.766,11.875 8.857,12.102 9.039,12.303C9.226,12.5 9.453,12.598 9.72,12.598ZM20.741,12.598H23.351C23.618,12.598 23.842,12.5 24.024,12.303C24.212,12.107 24.305,11.88 24.305,11.623C24.305,11.351 24.214,11.12 24.031,10.928C23.849,10.736 23.622,10.641 23.351,10.641H20.741C20.479,10.641 20.255,10.739 20.068,10.935C19.885,11.127 19.794,11.356 19.794,11.623C19.794,11.875 19.888,12.102 20.075,12.303C20.262,12.5 20.484,12.598 20.741,12.598ZM9.72,16.281H10.583C10.849,16.281 11.074,16.183 11.256,15.987C11.439,15.785 11.53,15.556 11.53,15.299C11.53,15.028 11.439,14.796 11.256,14.604C11.078,14.413 10.854,14.317 10.583,14.317H9.72C9.453,14.317 9.226,14.415 9.039,14.611C8.857,14.808 8.766,15.037 8.766,15.299C8.766,15.552 8.857,15.778 9.039,15.979C9.226,16.181 9.453,16.281 9.72,16.281ZM13.677,16.281H16.967C17.233,16.281 17.458,16.183 17.64,15.987C17.827,15.785 17.921,15.556 17.921,15.299C17.921,15.028 17.83,14.796 17.647,14.604C17.465,14.413 17.238,14.317 16.967,14.317H13.677C13.41,14.317 13.185,14.415 13.003,14.611C12.821,14.808 12.729,15.037 12.729,15.299C12.729,15.552 12.821,15.778 13.003,15.979C13.19,16.181 13.415,16.281 13.677,16.281ZM20.061,16.281H23.351C23.622,16.281 23.849,16.183 24.031,15.987C24.214,15.785 24.305,15.556 24.305,15.299C24.305,15.028 24.214,14.796 24.031,14.604C23.849,14.413 23.622,14.317 23.351,14.317H20.061C19.794,14.317 19.569,14.415 19.387,14.611C19.205,14.808 19.114,15.037 19.114,15.299C19.114,15.552 19.205,15.778 19.387,15.979C19.574,16.181 19.799,16.281 20.061,16.281ZM9.72,19.964H12.33C12.587,19.964 12.809,19.864 12.996,19.663C13.183,19.462 13.277,19.235 13.277,18.982C13.277,18.72 13.183,18.493 12.996,18.302C12.814,18.105 12.591,18.007 12.33,18.007H9.72C9.448,18.007 9.222,18.103 9.039,18.295C8.857,18.486 8.766,18.716 8.766,18.982C8.766,19.244 8.857,19.473 9.039,19.67C9.226,19.866 9.453,19.964 9.72,19.964ZM15.423,19.964H23.351C23.618,19.964 23.842,19.864 24.024,19.663C24.212,19.462 24.305,19.235 24.305,18.982C24.305,18.72 24.214,18.493 24.031,18.302C23.849,18.105 23.622,18.007 23.351,18.007H15.423C15.152,18.007 14.925,18.103 14.743,18.295C14.561,18.486 14.469,18.716 14.469,18.982C14.469,19.244 14.561,19.473 14.743,19.67C14.925,19.866 15.152,19.964 15.423,19.964ZM9.72,23.641H16.728C16.995,23.641 17.219,23.542 17.402,23.346C17.589,23.149 17.682,22.923 17.682,22.666C17.682,22.394 17.589,22.163 17.402,21.971C17.219,21.774 16.995,21.676 16.728,21.676H9.72C9.448,21.676 9.222,21.774 9.039,21.971C8.857,22.167 8.766,22.399 8.766,22.666C8.766,22.918 8.857,23.142 9.039,23.339C9.226,23.54 9.453,23.641 9.72,23.641Z"
+      android:strokeAlpha="0.5"
+      android:fillAlpha="0.5">
+    <aapt:attr name="android:fillColor">
+      <gradient 
+          android:startX="16.535"
+          android:startY="10.641"
+          android:endX="16.535"
+          android:endY="23.641"
+          android:type="linear">
+        <item android:offset="0" android:color="#FFFFFFFF"/>
+        <item android:offset="1" android:color="#FFF2F2F2"/>
+      </gradient>
+    </aapt:attr>
+  </path>
+  <path
+      android:pathData="M9.72,12.598H17.647C17.919,12.598 18.145,12.5 18.328,12.303C18.51,12.107 18.601,11.88 18.601,11.623C18.601,11.351 18.51,11.12 18.328,10.928C18.15,10.736 17.923,10.641 17.647,10.641H9.72C9.448,10.641 9.222,10.739 9.039,10.935C8.857,11.127 8.766,11.356 8.766,11.623C8.766,11.875 8.857,12.102 9.039,12.303C9.226,12.5 9.453,12.598 9.72,12.598ZM20.061,16.281H23.351C23.622,16.281 23.849,16.183 24.031,15.987C24.214,15.785 24.305,15.556 24.305,15.299C24.305,15.028 24.214,14.796 24.031,14.604C23.849,14.413 23.622,14.317 23.351,14.317H20.061C19.794,14.317 19.569,14.415 19.387,14.611C19.205,14.808 19.114,15.037 19.114,15.299C19.114,15.552 19.205,15.778 19.387,15.979C19.574,16.181 19.799,16.281 20.061,16.281ZM9.72,19.964H12.33C12.587,19.964 12.809,19.864 12.996,19.663C13.183,19.462 13.277,19.235 13.277,18.982C13.277,18.72 13.183,18.493 12.996,18.302C12.814,18.105 12.591,18.007 12.33,18.007H9.72C9.448,18.007 9.222,18.103 9.039,18.295C8.857,18.486 8.766,18.716 8.766,18.982C8.766,19.244 8.857,19.473 9.039,19.67C9.226,19.866 9.453,19.964 9.72,19.964ZM9.72,23.641H16.728C16.995,23.641 17.219,23.542 17.402,23.346C17.589,23.149 17.682,22.923 17.682,22.666C17.682,22.394 17.589,22.163 17.402,21.971C17.219,21.774 16.995,21.676 16.728,21.676H9.72C9.448,21.676 9.222,21.774 9.039,21.971C8.857,22.167 8.766,22.399 8.766,22.666C8.766,22.918 8.857,23.142 9.039,23.339C9.226,23.54 9.453,23.641 9.72,23.641Z">
+    <aapt:attr name="android:fillColor">
+      <gradient 
+          android:startX="16.535"
+          android:startY="10.641"
+          android:endX="16.535"
+          android:endY="23.641"
+          android:type="linear">
+        <item android:offset="0" android:color="#FFFFFFFF"/>
+        <item android:offset="1" android:color="#FFF2F2F2"/>
+      </gradient>
+    </aapt:attr>
+  </path>
+  <path
+      android:pathData="M24.69,25.775C24.84,25.925 25.022,26 25.237,26C25.452,26 25.635,25.925 25.784,25.775C25.934,25.619 26.009,25.433 26.009,25.218C26.016,25.003 25.941,24.821 25.784,24.671L9.329,8.225C9.179,8.069 8.993,7.994 8.772,8C8.55,8.007 8.368,8.082 8.225,8.225C8.075,8.375 8,8.561 8,8.782C8.007,9.004 8.081,9.186 8.225,9.329L24.69,25.775Z"
+      android:fillColor="#ffffff"/>
+  <path
+      android:pathData="M26.62,26.593L26.611,26.602C26.239,26.975 25.757,27.169 25.237,27.169C24.717,27.169 24.236,26.975 23.864,26.603C23.863,26.602 23.864,26.603 23.864,26.603L7.398,10.157C7.024,9.782 6.845,9.306 6.831,8.816L6.831,8.799V8.782C6.831,8.266 7.018,7.778 7.398,7.398C7.773,7.023 8.248,6.846 8.738,6.831C9.276,6.816 9.779,7.011 10.163,7.405L26.604,23.837C26.986,24.209 27.19,24.699 27.178,25.236C27.174,25.748 26.981,26.216 26.629,26.584L26.62,26.593ZM25.784,24.671L9.329,8.225C9.179,8.069 8.993,7.994 8.772,8C8.55,8.007 8.368,8.082 8.225,8.225C8.075,8.375 8,8.561 8,8.782C8.006,9.004 8.081,9.186 8.225,9.329L24.69,25.775C24.84,25.925 25.022,26 25.237,26C25.452,26 25.635,25.925 25.784,25.775C25.934,25.619 26.009,25.433 26.009,25.218C26.016,25.003 25.941,24.821 25.784,24.671Z"
+      android:fillColor="#78C4F5"
+      android:fillType="evenOdd"/>
+</vector>
