--- wallet/res/drawable/ic_key_filled.xml
@@ -1,21 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:aapt="http://schemas.android.com/aapt"
-    android:width="19dp"
-    android:height="36dp"
-    android:viewportWidth="19"
-    android:viewportHeight="36">
-  <path
-      android:pathData="M9.3201,0C8.0171,0 6.8041,0.2382 5.681,0.7147C4.5579,1.1911 3.571,1.8548 2.7202,2.7056C1.8694,3.5564 1.2033,4.5434 0.722,5.6664C0.2407,6.7895 0,7.9977 0,9.2909C0,10.5939 0.2455,11.8215 0.7366,12.9737C1.2276,14.126 1.9204,15.1372 2.815,16.0075C3.7095,16.8777 4.7646,17.5462 5.98,18.013V32.4088C5.98,32.5935 6.0116,32.7588 6.0748,32.9047C6.138,33.0505 6.2328,33.1915 6.3592,33.3276L8.795,35.778C8.9214,35.9044 9.0891,35.9773 9.2982,35.9968C9.5072,36.0162 9.6896,35.9481 9.8451,35.7926L14.4687,31.169C14.6243,31.0134 14.6996,30.8384 14.6948,30.6439C14.6899,30.4495 14.6146,30.2744 14.4687,30.1188L11.5954,27.2893L15.5334,23.3658C15.6793,23.2199 15.7498,23.0498 15.7449,22.8553C15.7401,22.6608 15.6598,22.4858 15.5043,22.3303L11.61,18.4068C13.8658,17.5122 15.5991,16.2676 16.8097,14.6729C18.0203,13.0782 18.6255,11.2842 18.6255,9.2909C18.6255,8.0074 18.3849,6.8041 17.9036,5.681C17.4222,4.5579 16.7562,3.571 15.9054,2.7202C15.0546,1.8694 14.0652,1.2033 12.9372,0.722C11.8093,0.2407 10.6036,0 9.3201,0ZM9.3201,8.9117C8.6005,8.9117 7.9879,8.6589 7.4823,8.1532C6.9767,7.6476 6.7238,7.035 6.7238,6.3155C6.7238,5.5959 6.9742,4.9833 7.475,4.4777C7.9758,3.9721 8.5908,3.7193 9.3201,3.7193C10.0299,3.7193 10.64,3.9745 11.1505,4.485C11.661,4.9955 11.9163,5.6056 11.9163,6.3155C11.9163,7.035 11.661,7.6476 11.1505,8.1532C10.64,8.6589 10.0299,8.9117 9.3201,8.9117Z">
-    <aapt:attr name="android:fillColor">
-      <gradient 
-          android:startY="0"
-          android:startX="9.31277"
-          android:endY="36"
-          android:endX="9.31277"
-          android:type="linear">
-        <item android:offset="0" android:color="#FF78C4F5"/>
-        <item android:offset="1" android:color="#FF008DE4"/>
-      </gradient>
-    </aapt:attr>
-  </path>
-</vector>
