+++ wallet/res/drawable/ic_no_restore.xml
@@ -1,0 +1,19 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="34dp"
+    android:height="34dp"
+    android:viewportWidth="34"
+    android:viewportHeight="34">
+  <path
+      android:pathData="M17,0L17,0A17,17 0,0 1,34 17L34,17A17,17 0,0 1,17 34L17,34A17,17 0,0 1,0 17L0,17A17,17 0,0 1,17 0z"
+      android:fillColor="#78C4F5"/>
+  <path
+      android:pathData="M17,25.696C18.189,25.696 19.307,25.468 20.352,25.012C21.403,24.561 22.329,23.936 23.13,23.135C23.93,22.334 24.559,21.411 25.015,20.366C25.471,19.315 25.699,18.192 25.699,16.997C25.699,15.807 25.471,14.689 25.015,13.644C24.559,12.593 23.93,11.667 23.13,10.866C22.329,10.066 21.403,9.44 20.352,8.99C19.307,8.534 18.189,8.306 17,8.306C15.81,8.306 14.692,8.531 13.647,8.981C12.601,9.426 11.678,10.046 10.877,10.841C10.077,11.636 9.449,12.557 8.993,13.602C8.537,14.647 8.306,15.762 8.3,16.947H10.21C10.216,16.007 10.394,15.131 10.744,14.319C11.094,13.502 11.581,12.787 12.204,12.176C12.826,11.564 13.547,11.086 14.364,10.741C15.181,10.391 16.06,10.216 17,10.216C17.939,10.216 18.818,10.391 19.635,10.741C20.458,11.091 21.178,11.578 21.795,12.201C22.418,12.818 22.904,13.538 23.255,14.361C23.611,15.178 23.789,16.057 23.789,16.997C23.789,17.942 23.611,18.826 23.255,19.649C22.904,20.466 22.418,21.186 21.795,21.809C21.178,22.432 20.458,22.918 19.635,23.269C18.818,23.619 17.939,23.791 17,23.786C16.282,23.786 15.601,23.683 14.956,23.477C14.317,23.271 13.724,22.979 13.179,22.601C12.64,22.218 12.162,21.767 11.745,21.25C11.55,21.039 11.325,20.908 11.069,20.858C10.814,20.803 10.575,20.864 10.352,21.042C10.141,21.203 10.024,21.417 10.002,21.684C9.985,21.945 10.085,22.204 10.302,22.459C10.802,23.099 11.406,23.663 12.112,24.153C12.824,24.642 13.597,25.02 14.431,25.287C15.27,25.559 16.126,25.696 17,25.696ZM7.308,15.237C7.08,15.237 6.907,15.29 6.791,15.395C6.68,15.501 6.627,15.637 6.632,15.804C6.638,15.965 6.702,16.129 6.824,16.296L8.959,19.215C9.104,19.41 9.276,19.507 9.476,19.507C9.677,19.501 9.846,19.404 9.985,19.215L12.129,16.288C12.245,16.121 12.304,15.957 12.304,15.795C12.309,15.634 12.257,15.501 12.145,15.395C12.034,15.29 11.865,15.237 11.637,15.237H7.308Z"
+      android:fillColor="#ffffff"/>
+  <path
+      android:pathData="M24.736,25.82C24.886,25.969 25.068,26.044 25.283,26.044C25.498,26.044 25.681,25.969 25.83,25.82C25.98,25.663 26.055,25.478 26.055,25.263C26.062,25.048 25.987,24.865 25.83,24.715L9.375,8.27C9.225,8.113 9.039,8.038 8.818,8.045C8.596,8.051 8.414,8.126 8.271,8.27C8.121,8.419 8.046,8.605 8.046,8.827C8.052,9.048 8.127,9.23 8.271,9.374L24.736,25.82Z"
+      android:fillColor="#ffffff"/>
+  <path
+      android:pathData="M26.666,26.638L26.657,26.646C26.285,27.019 25.803,27.214 25.283,27.214C24.763,27.214 24.282,27.019 23.91,26.647C23.909,26.647 23.91,26.647 23.91,26.647L7.444,10.201C7.069,9.826 6.891,9.35 6.877,8.861L6.876,8.844V8.827C6.876,8.311 7.064,7.823 7.444,7.443C7.818,7.068 8.294,6.89 8.783,6.876C9.322,6.86 9.825,7.055 10.208,7.449L26.65,23.881C27.032,24.253 27.236,24.744 27.224,25.28C27.22,25.792 27.027,26.261 26.675,26.629L26.666,26.638ZM25.83,24.715L9.375,8.27C9.225,8.113 9.039,8.038 8.818,8.045C8.596,8.051 8.414,8.126 8.271,8.27C8.121,8.419 8.046,8.605 8.046,8.827C8.052,9.048 8.127,9.23 8.271,9.374L24.736,25.82C24.886,25.969 25.068,26.044 25.283,26.044C25.498,26.044 25.681,25.969 25.83,25.82C25.98,25.663 26.055,25.478 26.055,25.263C26.062,25.048 25.987,24.865 25.83,24.715Z"
+      android:fillColor="#78C4F5"
+      android:fillType="evenOdd"/>
+</vector>
