+++ wallet/res/drawable/ic_protected.xml
@@ -1,0 +1,35 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:aapt="http://schemas.android.com/aapt"
+    android:width="80dp"
+    android:height="80dp"
+    android:viewportWidth="80"
+    android:viewportHeight="80">
+  <path
+      android:pathData="M19,0L61,0A19,19 0,0 1,80 19L80,61A19,19 0,0 1,61 80L19,80A19,19 0,0 1,0 61L0,19A19,19 0,0 1,19 0z">
+    <aapt:attr name="android:fillColor">
+      <gradient 
+          android:startX="0"
+          android:startY="0"
+          android:endX="80"
+          android:endY="80"
+          android:type="linear">
+        <item android:offset="0" android:color="#FF3CB878"/>
+        <item android:offset="1" android:color="#FF5DCB92"/>
+      </gradient>
+    </aapt:attr>
+  </path>
+  <path
+      android:pathData="M40,61C40.169,61 40.373,60.965 40.613,60.894C40.866,60.838 41.127,60.739 41.394,60.599C44.408,59.021 46.93,57.599 48.958,56.331C51,55.063 52.627,53.81 53.838,52.57C55.063,51.331 55.944,50 56.479,48.577C57.014,47.141 57.282,45.472 57.282,43.57V27.641C57.282,26.556 57.049,25.775 56.584,25.296C56.134,24.817 55.465,24.394 54.577,24.028C54.084,23.817 53.408,23.549 52.549,23.225C51.69,22.901 50.739,22.549 49.697,22.169C48.669,21.789 47.634,21.409 46.591,21.028C45.563,20.648 44.62,20.31 43.761,20.014C42.915,19.704 42.246,19.465 41.754,19.296C41.472,19.211 41.183,19.141 40.887,19.084C40.591,19.028 40.296,19 40,19C39.704,19 39.408,19.028 39.113,19.084C38.831,19.141 38.549,19.211 38.268,19.296C37.775,19.465 37.098,19.704 36.239,20.014C35.38,20.31 34.43,20.648 33.387,21.028C32.359,21.394 31.324,21.775 30.282,22.169C29.253,22.549 28.31,22.909 27.451,23.247C26.591,23.57 25.915,23.831 25.423,24.028C24.549,24.38 23.88,24.803 23.415,25.296C22.951,25.775 22.718,26.556 22.718,27.641V43.57C22.718,45.472 22.993,47.127 23.542,48.535C24.106,49.93 25,51.232 26.225,52.444C27.465,53.641 29.098,54.88 31.127,56.162C33.169,57.444 35.662,58.923 38.606,60.599C38.873,60.739 39.127,60.838 39.366,60.894C39.62,60.965 39.831,61 40,61ZM32.141,47.838V39.176C32.141,37.669 32.775,36.873 34.042,36.789V34.232C34.042,32.218 34.591,30.599 35.69,29.373C36.789,28.148 38.225,27.535 40,27.535C41.775,27.535 43.211,28.148 44.31,29.373C45.408,30.599 45.958,32.218 45.958,34.232V36.789C47.239,36.873 47.88,37.669 47.88,39.176V47.838C47.88,49.43 47.134,50.225 45.641,50.225H34.359C32.88,50.225 32.141,49.43 32.141,47.838ZM36.324,36.768H43.676V33.979C43.676,32.711 43.338,31.69 42.662,30.916C42,30.127 41.113,29.732 40,29.732C38.901,29.732 38.014,30.127 37.338,30.916C36.662,31.69 36.324,32.711 36.324,33.979V36.768Z">
+    <aapt:attr name="android:fillColor">
+      <gradient 
+          android:startX="31.023"
+          android:startY="31.25"
+          android:endX="48.398"
+          android:endY="60.881"
+          android:type="linear">
+        <item android:offset="0" android:color="#FFFFFFFF"/>
+        <item android:offset="1" android:color="#FFF0F0F0"/>
+      </gradient>
+    </aapt:attr>
+  </path>
+</vector>
