--- wallet/res/drawable/ic_shield_lock_filled.xml
@@ -1,24 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:aapt="http://schemas.android.com/aapt"
-    android:width="30dp"
-    android:height="36dp"
-    android:viewportWidth="30"
-    android:viewportHeight="36">
-  <path
-      android:pathData="M14.8129,36C14.9578,36 15.1358,35.9728 15.3471,35.9185C15.5584,35.8642 15.7787,35.7767 16.0081,35.6559C18.5916,34.3038 20.7555,33.0845 22.5,31.998C24.2445,30.9115 25.6388,29.837 26.6831,28.7746C27.7274,27.7123 28.4789,26.5684 28.9376,25.343C29.3964,24.1177 29.6258,22.6901 29.6258,21.0604V7.4064C29.6258,6.4769 29.4296,5.8068 29.0372,5.3964C28.6449,4.9859 28.0684,4.6237 27.3078,4.3099C26.8853,4.1288 26.3059,3.8994 25.5694,3.6217C24.833,3.3441 24.0211,3.0423 23.1338,2.7163C22.2465,2.3903 21.3592,2.0644 20.4718,1.7384C19.5845,1.4125 18.7757,1.1197 18.0453,0.8601C17.3149,0.6006 16.7385,0.3984 16.3159,0.2535C16.0745,0.1811 15.827,0.1207 15.5734,0.0724C15.3199,0.0241 15.0664,0 14.8129,0C14.5594,0 14.3089,0.0241 14.0614,0.0724C13.8139,0.1207 13.5694,0.1811 13.328,0.2535C12.9054,0.3984 12.326,0.6006 11.5895,0.8601C10.8531,1.1197 10.0412,1.4095 9.1539,1.7294C8.2666,2.0493 7.3793,2.3752 6.492,2.7072C5.6046,3.0392 4.7927,3.3471 4.0563,3.6308C3.3199,3.9145 2.7404,4.1408 2.3179,4.3099C1.5694,4.6117 0.996,4.9708 0.5976,5.3873C0.1992,5.8038 0,6.4769 0,7.4064V21.0604C0,22.6901 0.2384,24.1056 0.7153,25.3068C1.1922,26.508 1.9588,27.6247 3.0151,28.6569C4.0714,29.6891 5.4718,30.7545 7.2163,31.8531C8.9608,32.9517 11.0946,34.2193 13.6177,35.6559C13.8471,35.7767 14.0674,35.8642 14.2787,35.9185C14.49,35.9728 14.668,36 14.8129,36Z">
-    <aapt:attr name="android:fillColor">
-      <gradient 
-          android:startY="0"
-          android:startX="14.8129"
-          android:endY="36"
-          android:endX="14.8129"
-          android:type="linear">
-        <item android:offset="0" android:color="#FF78C4F5"/>
-        <item android:offset="1" android:color="#FF008DE4"/>
-      </gradient>
-    </aapt:attr>
-  </path>
-  <path
-      android:pathData="M8.0764,24.7189V17.2944C8.0764,16.0026 8.6197,15.3205 9.7062,15.2481V13.057C9.7062,11.3306 10.177,9.9422 11.1187,8.8919C12.0603,7.8417 13.2917,7.3165 14.8128,7.3165C16.334,7.3165 17.5653,7.8417 18.507,8.8919C19.4487,9.9422 19.9195,11.3306 19.9195,13.057V15.2481C21.0181,15.3205 21.5673,16.0026 21.5673,17.2944V24.7189C21.5673,26.0831 20.9275,26.7652 19.6478,26.7652H9.9778C8.7102,26.7652 8.0764,26.0831 8.0764,24.7189ZM11.6619,15.23H17.9637V12.8396C17.9637,11.7531 17.677,10.8748 17.1036,10.2048C16.5301,9.5348 15.7666,9.1998 14.8128,9.1998C13.8712,9.1998 13.1106,9.5348 12.5311,10.2048C11.9516,10.8748 11.6619,11.7531 11.6619,12.8396V15.23Z"
-      android:fillColor="#ffffff"/>
-</vector>
