--- wallet/res/drawable/ic_square_stack.xml
@@ -1,35 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:aapt="http://schemas.android.com/aapt"
-    android:width="33dp"
-    android:height="34dp"
-    android:viewportWidth="33"
-    android:viewportHeight="34">
-  <path
-      android:pathData="M10.9351,16.2955L2.6342,7.9946C1.8583,8.5255 1.4703,9.1075 1.4703,9.7405C1.4703,10.1387 1.5928,10.4884 1.8378,10.7896C2.0829,11.0908 2.5015,11.415 3.0937,11.7622L10.9351,16.2955ZM21.1351,17.582L31.0594,11.8234C31.6823,11.4558 32.124,11.1164 32.3843,10.805C32.6446,10.4935 32.7748,10.1234 32.7748,9.6946C32.7748,9.276 32.6446,8.911 32.3843,8.5996C32.124,8.2881 31.6823,7.9487 31.0594,7.5811L19.0829,0.6279C18.7153,0.4135 18.3758,0.2553 18.0644,0.1532C17.753,0.0511 17.4441,0 17.1378,0C16.8213,0 16.5099,0.0511 16.2036,0.1532C15.8973,0.2553 15.5553,0.4135 15.1775,0.6279L8.2243,4.6559L21.1351,17.582ZM17.1378,26.9549C17.4135,26.9549 17.679,26.8937 17.9342,26.7712C18.1895,26.6486 18.4805,26.4955 18.8072,26.3117L20.4,25.3622L18.5928,23.555L17.7198,24.0604C17.4646,24.2033 17.2706,24.2748 17.1378,24.2748C16.9949,24.2748 16.7958,24.2033 16.5405,24.0604L2.3586,15.8667C2.1441,15.9688 1.9425,16.1398 1.7536,16.3797C1.5647,16.6197 1.4703,16.8979 1.4703,17.2144C1.4703,17.4492 1.5392,17.6866 1.677,17.9266C1.8149,18.1665 2.0268,18.3733 2.3126,18.5469L15.4532,26.3117C15.7799,26.4955 16.0709,26.6486 16.3261,26.7712C16.5814,26.8937 16.852,26.9549 17.1378,26.9549ZM25.745,22.2072L31.9631,18.5469C32.2489,18.3733 32.4582,18.1665 32.591,17.9266C32.7238,17.6866 32.7902,17.4492 32.7902,17.2144C32.7902,16.8979 32.6957,16.6197 32.5066,16.3797C32.3178,16.1398 32.1162,15.9688 31.9019,15.8667L23.9838,20.4306L25.745,22.2072ZM17.1378,34C17.4135,34 17.679,33.9413 17.9342,33.8239C18.1895,33.7064 18.4805,33.5507 18.8072,33.3568L24.8414,29.8036L23.0495,28.027L17.7198,31.1207C17.6075,31.182 17.5028,31.233 17.4059,31.2739C17.3089,31.3147 17.2195,31.3351 17.1378,31.3351C17.0459,31.3351 16.9515,31.3147 16.8545,31.2739C16.7575,31.233 16.6529,31.182 16.5405,31.1207L2.3586,22.9117C2.1441,23.024 1.9425,23.2001 1.7536,23.4401C1.5647,23.68 1.4703,23.9531 1.4703,24.2595C1.4703,24.5045 1.5392,24.747 1.677,24.9869C1.8149,25.2269 2.0268,25.4285 2.3126,25.5919L15.4532,33.3568C15.7799,33.5507 16.0709,33.7064 16.3261,33.8239C16.5814,33.9413 16.852,34 17.1378,34ZM30.1865,26.6486L31.9631,25.5919C32.2387,25.4285 32.4455,25.2269 32.5835,24.9869C32.7213,24.747 32.7902,24.5045 32.7902,24.2595C32.7902,23.9531 32.6957,23.68 32.5066,23.4401C32.3178,23.2001 32.1162,23.024 31.9019,22.9117L28.4559,24.918L30.1865,26.6486Z">
-    <aapt:attr name="android:fillColor">
-      <gradient 
-          android:startY="0"
-          android:startX="17.1302"
-          android:endY="34"
-          android:endX="17.1302"
-          android:type="linear">
-        <item android:offset="0" android:color="#FF78C4F5"/>
-        <item android:offset="1" android:color="#FF008DE4"/>
-      </gradient>
-    </aapt:attr>
-  </path>
-  <path
-      android:pathData="M30.7837,33.1422C30.9982,33.3668 31.2739,33.4791 31.6108,33.4791C31.9477,33.4791 32.2285,33.3668 32.4533,33.1422C32.6779,32.9073 32.7902,32.6266 32.7902,32.2998C32.7902,31.9731 32.6779,31.6974 32.4533,31.4728L2.0216,1.0566C1.7868,0.832 1.506,0.7171 1.1793,0.712C0.8526,0.7069 0.5718,0.8218 0.3369,1.0566C0.1123,1.2813 0,1.5595 0,1.8913C0,2.2231 0.1123,2.5014 0.3369,2.726L30.7837,33.1422Z">
-    <aapt:attr name="android:fillColor">
-      <gradient 
-          android:startY="0.711855"
-          android:startX="16.3951"
-          android:endY="33.4791"
-          android:endX="16.3951"
-          android:type="linear">
-        <item android:offset="0" android:color="#FF78C4F5"/>
-        <item android:offset="1" android:color="#FF008DE4"/>
-      </gradient>
-    </aapt:attr>
-  </path>
-</vector>
