+++ wallet/res/layout/dialog_verify_seed_warning.xml
@@ -7,7 +7,7 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:paddingBottom="18dp"
-    tools:context="de.schildbach.wallet.ui.VerifySeedWarningDialog">
+    tools:context="de.schildbach.wallet.ui.verify.VerifySeedWarningDialog">
 
     <ImageButton
         android:id="@+id/collapse_button"
