--- wallet/res/layout/fragment_its_import_to_secure.xml
@@ -1,143 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="de.schildbach.wallet.ui.VerifySeedItIsImportantFragment">
-
-    <include
-        android:id="@+id/app_bar"
-        layout="@layout/app_bar_general" />
-
-    <ScrollView
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:clipChildren="false"
-        android:orientation="vertical"
-        android:paddingBottom="20dp"
-        android:clipToPadding="false"
-        app:layout_constraintBottom_toTopOf="@+id/verify_show_recovery_phrase_button"
-        app:layout_constraintTop_toBottomOf="@+id/app_bar">
-
-        <androidx.constraintlayout.widget.ConstraintLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content">
-
-            <TextView
-                android:id="@+id/title"
-                style="@style/Headline5"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="36dp"
-                android:text="@string/verify_it_is_important"
-                app:layout_constraintTop_toTopOf="parent"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintEnd_toEndOf="parent"/>
-
-            <TextView
-                android:id="@+id/textView"
-                style="@style/Body2"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="4dp"
-                android:layout_marginHorizontal="25dp"
-                android:lineSpacingExtra="6dp"
-                android:gravity="center"
-                android:textAlignment="gravity"
-                android:text="@string/verify_it_is_important_subtitle"
-                app:layout_constraintTop_toBottomOf="@+id/title"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintEnd_toEndOf="parent" />
-
-            <ImageView
-                android:id="@+id/info_icon_1"
-                android:layout_width="50dp"
-                android:layout_height="50dp"
-                android:layout_marginStart="20dp"
-                android:layout_marginTop="32dp"
-                android:padding="7dp"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toBottomOf="@id/textView"
-                app:srcCompat="@drawable/ic_key_filled" />
-
-            <TextView
-                android:id="@+id/info_text_1"
-                style="@style/Caption"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:minHeight="50dp"
-                android:layout_marginStart="15dp"
-                android:layout_marginEnd="50dp"
-                android:layout_marginTop="6dp"
-                android:lineSpacingExtra="4dp"
-                android:text="@string/verify_info_text_1"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintStart_toEndOf="@id/info_icon_1"
-                app:layout_constraintTop_toTopOf="@id/info_icon_1" />
-
-            <ImageView
-                android:id="@+id/info_icon_2"
-                android:layout_width="50dp"
-                android:layout_height="50dp"
-                android:layout_marginTop="14dp"
-                android:padding="7dp"
-                app:layout_constraintStart_toStartOf="@+id/info_icon_1"
-                app:layout_constraintTop_toBottomOf="@id/info_text_1"
-                app:srcCompat="@drawable/ic_square_stack" />
-
-            <TextView
-                android:id="@+id/info_text_2"
-                style="@style/Caption"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:minHeight="50dp"
-                android:layout_marginStart="15dp"
-                android:layout_marginEnd="50dp"
-                android:layout_marginTop="6dp"
-                android:lineSpacingExtra="4dp"
-                android:text="@string/verify_info_text_2"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintStart_toEndOf="@+id/info_icon_2"
-                app:layout_constraintTop_toTopOf="@+id/info_icon_2" />
-
-            <ImageView
-                android:id="@+id/info_icon_3"
-                android:layout_width="50dp"
-                android:layout_height="50dp"
-                android:padding="7dp"
-                android:layout_marginTop="14dp"
-                app:layout_constraintStart_toStartOf="@+id/info_icon_2"
-                app:layout_constraintTop_toBottomOf="@id/info_text_2"
-                app:srcCompat="@drawable/ic_shield_lock_filled" />
-
-            <TextView
-                android:id="@+id/info_text_3"
-                style="@style/Caption"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:minHeight="50dp"
-                android:layout_marginStart="15dp"
-                android:layout_marginEnd="50dp"
-                android:layout_marginTop="6dp"
-                android:lineSpacingExtra="4dp"
-                android:text="@string/verify_info_text_3"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintStart_toEndOf="@+id/info_icon_3"
-                app:layout_constraintTop_toTopOf="@+id/info_icon_3" />
-        </androidx.constraintlayout.widget.ConstraintLayout>
-    </ScrollView>
-
-    <Button
-        android:id="@+id/show_recovery_phrase_button"
-        style="@style/Button.Primary.Large.Blue"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginHorizontal="16dp"
-        android:layout_marginBottom="16dp"
-        android:text="@string/verify_show_recovery_phrase"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent" />
-</androidx.constraintlayout.widget.ConstraintLayout>
