+++ wallet/res/layout/fragment_verify_confirm.xml
@@ -1,11 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/linearLayout2"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/background_primary"
-    android:gravity="center_horizontal">
+    android:gravity="center_horizontal"
+    tools:context="de.schildbach.wallet.ui.verify.VerifySeedConfirmFragment">
 
     <include
         android:id="@+id/verify_appbar"
