+++ wallet/res/values/styles.xml
@@ -41,6 +41,7 @@
         <item name="android:background">@drawable/rounded_background</item>
         <item name="android:theme">@style/VerifySeedTextViewBackground</item>
         <item name="android:gravity">center</item>
+        <item name="android:textAlignment">gravity</item>
         <item name="android:lineSpacingMultiplier">1.7</item>
         <item name="android:paddingTop">18dp</item>
         <item name="android:paddingBottom">24dp</item>
