+++ wallet/src/de/schildbach/wallet/ui/SetPinActivity.kt
@@ -34,6 +34,7 @@ import de.schildbach.wallet.livedata.Status
 import de.schildbach.wallet.security.SecurityFunctions
 import de.schildbach.wallet.service.RestartService
 import de.schildbach.wallet.ui.main.WalletActivity
+import de.schildbach.wallet.ui.verify.VerifySeedActivity
 import de.schildbach.wallet.ui.widget.PinPreviewView
 import de.schildbach.wallet_test.R
 import kotlinx.coroutines.launch
