+++ wallet/src/de/schildbach/wallet/ui/verify/ViewSeedActivity.kt
@@ -1,32 +1,32 @@
 /*
- * Copyright 2019 Dash Core Group
+ * Copyright 2023 Dash Core Group.
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
  *
- *    http://www.apache.org/licenses/LICENSE-2.0
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package de.schildbach.wallet.ui
+package de.schildbach.wallet.ui.verify
 
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
-import dagger.hilt.android.AndroidEntryPoint
+import de.schildbach.wallet.ui.BaseMenuActivity
 import de.schildbach.wallet_test.R
 import kotlinx.android.synthetic.main.activity_view_seed.*
 
 /**
  * @author Eric Britten
  */
-@AndroidEntryPoint
 class ViewSeedActivity : BaseMenuActivity() {
 
     companion object {
@@ -53,8 +53,10 @@ class ViewSeedActivity : BaseMenuActivity() {
         seed = if (intent.extras?.containsKey(EXTRA_SEED)!!) {
             intent.extras!!.getStringArray(EXTRA_SEED)!!
         } else {
-            throw IllegalStateException("This activity needs to receive a String[] Intent Extra " +
-                    "containing the recovery seed.")
+            throw IllegalStateException(
+                "This activity needs to receive a String[] Intent Extra " +
+                    "containing the recovery seed."
+            )
         }
 
         val sb = seed.joinToString(" ")
@@ -62,7 +64,7 @@ class ViewSeedActivity : BaseMenuActivity() {
         recovery_seed.text = sb.trim()
 
         confirm_btn.setOnClickListener {
-           finish()
+            finish()
         }
 
         explanation_btn.setOnClickListener {
